Revert "ddl: fix race in table lock config (#10848)"
This reverts commit e4204df780bcb63fa211418b77191ad0dbba9898.
