+++ ddl/db_test.go
@@ -89,8 +89,6 @@ func setUpSuite(s *testDBSuite, c *C) {
 	s.schemaName = "test_db"
 	s.autoIDStep = autoid.GetStep()
 	ddl.WaitTimeWhenErrorOccured = 0
-	// Test for table lock.
-	config.GetGlobalConfig().EnableTableLock = true
 
 	s.cluster = mocktikv.NewCluster()
 	mocktikv.BootstrapWithSingleStore(s.cluster)
@@ -2778,6 +2776,18 @@ func (s *testDBSuite2) TestLockTables(c *C) {
 	tk.MustExec("create table t1 (a int)")
 	tk.MustExec("create table t2 (a int)")
 
+	// recover table lock config.
+	originValue := config.GetGlobalConfig().EnableTableLock
+	defer func() {
+		config.GetGlobalConfig().EnableTableLock = originValue
+	}()
+
+	// Test for enable table lock config.
+	config.GetGlobalConfig().EnableTableLock = false
+	tk.MustExec("lock tables t1 write")
+	checkTableLock(c, tk.Se, "test", "t1", model.TableLockNone)
+	config.GetGlobalConfig().EnableTableLock = true
+
 	// Test lock 1 table.
 	tk.MustExec("lock tables t1 write")
 	checkTableLock(c, tk.Se, "test", "t1", model.TableLockWrite)
@@ -2953,7 +2963,7 @@ func (s *testDBSuite2) TestLockTables(c *C) {
 }
 
 // TestConcurrentLockTables test concurrent lock/unlock tables.
-func (s *testDBSuite4) TestConcurrentLockTables(c *C) {
+func (s *testDBSuite2) TestConcurrentLockTables(c *C) {
 	if israce.RaceEnabled {
 		c.Skip("skip race test")
 	}
@@ -2965,6 +2975,15 @@ func (s *testDBSuite4) TestConcurrentLockTables(c *C) {
 	defer tk.MustExec("drop table if exists t1")
 	tk.MustExec("create table t1 (a int)")
 	tk2.MustExec("use test")
+
+	// recover table lock config.
+	originValue := config.GetGlobalConfig().EnableTableLock
+	defer func() {
+		config.GetGlobalConfig().EnableTableLock = originValue
+	}()
+
+	// Test for enable table lock config.
+	config.GetGlobalConfig().EnableTableLock = true
 
 	// Test concurrent lock tables read.
 	sql1 := "lock tables t1 read"
@@ -2998,7 +3017,7 @@ func (s *testDBSuite4) TestConcurrentLockTables(c *C) {
 	tk2.MustExec("unlock tables")
 }
 
-func (s *testDBSuite4) testParallelExecSQL(c *C, sql1, sql2 string, se1, se2 session.Session, f checkRet) {
+func (s *testDBSuite2) testParallelExecSQL(c *C, sql1, sql2 string, se1, se2 session.Session, f checkRet) {
 	callback := &ddl.TestDDLCallback{}
 	times := 0
 	callback.OnJobRunBeforeExported = func(job *model.Job) {
